Made lots of changes.
Started working towards making the graph 3D.
The perspective will need to be set at the creation of the context and every time the 'view' is changed.

More than one thread are now doing work.
There is now a mutex that protects the graph from being overwritten while being created.
	- This should probably be changed to lower level resource protection.
Program now locks on exit due to a race condition.